Move endpoint package into bulkscan package
