+++ src/main/java/uk/gov/hmcts/divorce/bulkscan/endpoint/ValidationController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.divorce.endpoint;
+package uk.gov.hmcts.divorce.bulkscan.endpoint;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -8,17 +8,17 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriUtils;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.data.FormType;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.data.OcrDataValidationRequest;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.data.OcrValidationResponse;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.data.ValidationStatus;
 import uk.gov.hmcts.divorce.bulkscan.validation.OcrValidator;
-import uk.gov.hmcts.divorce.endpoint.data.FormType;
-import uk.gov.hmcts.divorce.endpoint.data.OcrDataValidationRequest;
-import uk.gov.hmcts.divorce.endpoint.data.OcrValidationResponse;
 
 import java.nio.charset.StandardCharsets;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static org.apache.commons.lang3.EnumUtils.isValidEnum;
-import static uk.gov.hmcts.divorce.endpoint.data.ValidationStatus.ERRORS;
 
 @RestController
 public class ValidationController {
@@ -39,7 +39,7 @@ public class ValidationController {
             return ResponseEntity.ok().body(new OcrValidationResponse(
                 emptyList(),
                 singletonList("Form type '" + encodedFormType + "' not found"),
-                ERRORS
+                ValidationStatus.ERRORS
             ));
         }
