+++ src/main/java/uk/gov/hmcts/divorce/bulkscan/endpoint/data/ValidationStatus.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.divorce.endpoint.data;
+package uk.gov.hmcts.divorce.bulkscan.endpoint.data;
 
 import java.util.List;
