+++ src/main/java/uk/gov/hmcts/divorce/bulkscan/endpoint/model/input/TransformationInput.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.divorce.endpoint.model.input;
+package uk.gov.hmcts.divorce.bulkscan.endpoint.model.input;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
