+++ src/main/java/uk/gov/hmcts/divorce/bulkscan/transformation/BulkScanService.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.divorce.bulkscan.transformation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.divorce.endpoint.model.input.TransformationInput;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.model.input.TransformationInput;
 
 import java.util.Map;
