+++ src/main/java/uk/gov/hmcts/divorce/common/config/advice/GlobalExceptionHandler.java
@@ -8,11 +8,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.model.output.BspErrorResponse;
 import uk.gov.hmcts.divorce.common.config.interceptors.UnAuthorisedServiceException;
 import uk.gov.hmcts.divorce.common.exception.InvalidDataException;
 import uk.gov.hmcts.divorce.common.exception.UnsupportedFormTypeException;
 import uk.gov.hmcts.divorce.document.print.exception.InvalidResourceException;
-import uk.gov.hmcts.divorce.endpoint.model.output.BspErrorResponse;
 import uk.gov.hmcts.divorce.notification.exception.NotificationException;
 import uk.gov.hmcts.reform.authorisation.exceptions.InvalidTokenException;
 import uk.gov.service.notify.NotificationClientException;
