+++ src/test/java/uk/gov/hmcts/divorce/bulkscan/transformation/D8sFormToCaseTransformerTest.java
@@ -10,11 +10,11 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.data.OcrValidationResponse;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.model.input.OcrDataField;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.model.input.TransformationInput;
 import uk.gov.hmcts.divorce.common.exception.InvalidDataException;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.endpoint.data.OcrValidationResponse;
-import uk.gov.hmcts.divorce.endpoint.model.input.OcrDataField;
-import uk.gov.hmcts.divorce.endpoint.model.input.TransformationInput;
 
 import java.util.HashMap;
 import java.util.List;
@@ -32,14 +32,14 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.divorce.bulkscan.endpoint.data.FormType.D8;
+import static uk.gov.hmcts.divorce.bulkscan.endpoint.data.FormType.D8S;
 import static uk.gov.hmcts.divorce.bulkscan.transformation.CommonFormToCaseTransformer.TRANSFORMATION_AND_OCR_WARNINGS;
 import static uk.gov.hmcts.divorce.bulkscan.util.FileUtil.loadJson;
 import static uk.gov.hmcts.divorce.bulkscan.validation.data.OcrDataFields.transformOcrMapToObject;
 import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.SOLE_APPLICATION;
 import static uk.gov.hmcts.divorce.divorcecase.model.DivorceOrDissolution.DIVORCE;
 import static uk.gov.hmcts.divorce.divorcecase.model.SupplementaryCaseType.JUDICIAL_SEPARATION;
-import static uk.gov.hmcts.divorce.endpoint.data.FormType.D8;
-import static uk.gov.hmcts.divorce.endpoint.data.FormType.D8S;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.inputScannedDocuments;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.scannedDocuments;
