+++ src/test/java/uk/gov/hmcts/divorce/bulkscan/transformation/PaperFormDetailsTransformerTest.java
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.divorce.bulkscan.endpoint.model.input.OcrDataField;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.PaperFormDetails;
-import uk.gov.hmcts.divorce.endpoint.model.input.OcrDataField;
 
 import java.util.List;
